auto defend protocol

+++ src/common/mqtt.h
@@ -19,7 +19,7 @@ typedef struct
 void mqtt_initial(MQTT_ARG* mqtt_arg);
 void mqtt_cleanup();
 
-void mqtt_publish(const char *topic, const void *payload, int payloadlen);
+void mqtt_publish(const char *topic, const void *payload, size_t payloadlen);
 void mqtt_subscribe(const char *imei);
 void mqtt_unsubscribe(const char *imei);
